requirements.txt
fastapi
uvicorn
elasticsearch
sentence-transformers
pillow
torch